moved the lowDB transport to it's own file - it seems very specific to the bus app
