+++ lib/index.js
@@ -22,7 +22,7 @@ var transporter = nodemailer.createTransport({
 var crypto = require('crypto');
 var rollbar = require('rollbar');
 rollbar.init(config.ROLLBAR_TOKEN);
-var logger = require('./logTransport').logger
+var logger = require('./logger')
 
 
 /*
