+++ lib/logger.js
@@ -5,10 +5,6 @@ var onFinished = require('on-finished')
 var util = require('util');
 var onHeaders = require('on-headers');
 var hashwords = require('hashwords')();
-var low = require('lowdb');
-var db = low('./public/db.json', { storage: require('lowdb/lib/file-async') });
-var db_private = low('./db_private.json', { storage: require('lowdb/lib/file-async') });
-var moment = require('moment-timezone');
 var UUID = require("pure-uuid");
 
 var CONSOLE_LOG_LEVEL = 'debug';
@@ -37,7 +33,7 @@ logger.add(winston.transports.Console, {
 function consoleFormatter(options){
     // Send messages if they exist and also send formatted logger that logs all requests to info
     // Messages will usually be debug or warnings (to send message to console use logger.debug(message) ) 
-    // We only want custom format for the auto-logging. When called with logger.debug("message", args) it should
+    // We only want custom format for the auto-middleware logging. When called with logger.debug("message", args) it should
     // act like console.log but with a colorized level indication
     var meta = options.meta;
     var logMessage = (meta && 'status' in meta) 
@@ -47,82 +43,7 @@ function consoleFormatter(options){
     return logMessage
 }
 
-/* 
-    lowDB LOGGING
-    This writes local json formatted logs that can be querried using using lodash operators
-    It reads the entire file into memory so or reads and writes to stay in sync you need to
-    use the same handle. 
 
-*/
-var lowdbTransport = function (options){
-    this.name = "Local-Logs";
-    this.level = options.level || 'info';
-}
-
-util.inherits(lowdbTransport, winston.Transport);
-
-lowdbTransport.prototype.log = function(level, msg, meta, callback) {
-    if (level != 'info') return; // don't send warnins and debug messages to analytics
-    var entry = {
-        date            : new Date(), 
-        totalTime       : meta.responseTime,
-        input           : meta.input, 
-        phone           : meta.phone,
-        ip              : meta.ip,
-        stop            : meta.stopId,
-        geocodedAddress : meta.geocodedAddress,
-        muniTime        : meta.muniTime,
-        geocodeTime     : meta.geocodeTime
-    }
-    db_private.defaults({ requests: []}).value()
-    db_private.get('requests').push(entry).value();
-    var entry2 = JSON.parse(JSON.stringify(entry)); // Required because of async mode of lowdb
-    if (entry.phone) {
-        entry2.phone = hashwords.hashStr(entry.phone);
-    }
-    entry2.ip = "";
-    db.defaults({ requests: []}).value()
-    db.get('requests').push(entry2).value(); 
-}
-logger.add(lowdbTransport, {})
-
-//     NOTE: this code was moved out of routes/index.js into getLogData so we could use the same 
-//    handler from low('./db_private.json') without having to export just the handlerIf you call
-//    low() in two different places you get two different in-memory versions of the same File.
-
-function getLogData(daysBack, type){
-    var logData = [];
-    var timezone = moment.tz.zone(config.TIMEZONE);
-    if (type == "hits") {      
-        var dateTz = null;
-        db_private.get('requests').filter(function(point) {
-            if (point.date) {
-                var nowTz = moment.tz(new Date(), config.TIMEZONE);
-                dateTz = moment.tz(point.date, config.TIMEZONE);
-                if (moment.duration(nowTz.diff(dateTz)).asDays() <= daysBack) {
-                    return true;
-                }
-            }
-            return false;
-        }).forEach(function(point) {
-            var hitType = "browser";
-            if (point.phone) {
-                hitType = "sms"
-            }
-            var outPoint = {};
-            outPoint.type = hitType;
-            outPoint.date = moment.tz(point.date, config.TIMEZONE).unix();
-            outPoint.dateOffset = timezone.offset(outPoint.date*1000)
-            outPoint.muniTime = point.muniTime || "";
-            outPoint.totalTime = point.totalTime || "";
-            outPoint.userId = point.phone ? "phone" + point.phone : "ip"+point.ip;
-            logData.push(outPoint);
-        }).value()
-    }
-    return logData;
-}
-
-  
 /* 
     GOOGLE ANALYTICS LOG TRANSPORT
     This transport sends events to google analytics which can be tracked independently of pageviews
@@ -166,7 +87,7 @@ function initGoogleAnalytics(initFunction){
 
         var trackingCode = _fields.trackingCode
         if (!trackingCode) return logger.warn("Google Analytics requires a tracking code")
-        var visitor = ua(trackingCode, uuid, {strictCidFormat: false}).debug(); //strictCidFormat so GA accepts v5 UUID 
+        var visitor = ua(trackingCode, uuid, {strictCidFormat: false}); //strictCidFormat so GA accepts v5 UUID 
         visitor.set("uid",uuid)
 
         if (!_fields.action) { //Assume plain web hit if no action. At the moment this should only be '/'  and 404s 
@@ -187,7 +108,7 @@ function initGoogleAnalytics(initFunction){
    
         visitor.event(params, (err) => {
             if (err) logger.warn("google analytics event error: ", err)
-            })
+        })
         visitor.timing('Response Time', 'Total Time', meta.responseTime )
         if (_fields.timings){
                 if (!Array.isArray(_fields.timings)){ 
@@ -199,20 +120,21 @@ function initGoogleAnalytics(initFunction){
         }
         visitor.send()
     }
-
     logger.add(GATransport, {})
 }
+
+
 /*
     SETUP LOGGING MIDDLEWARE
     This will accept a function with a signature:
     initFunction(req, res) 
-    and it will return a middleware function.
-    The initFunction will be passsed the request and response object when the request
-    is finished. It should return an object that defines which the names and values
+    and will return a middleware function for use in app.use().
+    The initFunction will be passsed the request and response objects.
+    It should return an object that defines which the names and values
     of what should be logged for example:
     {phonenumber: req.body.From}
     These fields will be merged with the default fields. It is possible to 
-    override default fields by passing the same object key
+    override default fields by passing a value with the same object key as a defaul
 */
 function initialize(initFunction){
     return function log(req, res, next){
@@ -253,7 +175,7 @@ function getResponseTime(req, res){
          (res._startTime[1] - req._startTime[1]) * 1e-6;
     return ms.toFixed(0);
 }
-module.exports.logger = logger;
+
+module.exports = logger;
 module.exports.initialize = initialize;
-module.exports.getLogData = getLogData;
 module.exports.initGoogleAnalytics = initGoogleAnalytics;
